+++ drivers/video/mt9m114.c
@@ -16,8 +16,7 @@
 #include <zephyr/video/controls.h>
 #include <zephyr/video/formats.h>
 
-#include "video_ctrls.h"
-#include "video_device.h"
+#include "video_common.h"
 
 LOG_MODULE_REGISTER(video_mt9m114, CONFIG_VIDEO_LOG_LEVEL);
