+++ drivers/video/ov2640.c
@@ -15,8 +15,7 @@
 #include <zephyr/video/controls.h>
 #include <zephyr/video/formats.h>
 
-#include "video_ctrls.h"
-#include "video_device.h"
+#include "video_common.h"
 
 LOG_MODULE_REGISTER(video_ov2640, CONFIG_VIDEO_LOG_LEVEL);
