+++ drivers/video/ov9655.c
@@ -13,7 +13,6 @@
 #include <zephyr/video/formats.h>
 
 #include "video_common.h"
-#include "video_device.h"
 
 LOG_MODULE_REGISTER(video_ov9655, CONFIG_VIDEO_LOG_LEVEL);
