--- drivers/video/video_device.c
@@ -1,22 +1,0 @@
-/*
- * Copyright 2025 NXP
- *
- * SPDX-License-Identifier: Apache-2.0
- */
-
-#include "video_device.h"
-
-struct video_device *video_find_vdev(const struct device *dev)
-{
-	if (!dev) {
-		return NULL;
-	}
-
-	STRUCT_SECTION_FOREACH(video_device, vdev) {
-		if (vdev->dev == dev) {
-			return vdev;
-		}
-	}
-
-	return NULL;
-}
