+++ drivers/video/video_esp32_dvp.c
@@ -20,7 +20,7 @@
 
 #include <zephyr/logging/log.h>
 
-#include "video_device.h"
+#include "video_common.h"
 
 LOG_MODULE_REGISTER(video_esp32_lcd_cam, CONFIG_VIDEO_LOG_LEVEL);
