+++ drivers/video/video_st_mipid02.c
@@ -16,7 +16,6 @@
 #include <zephyr/video/video.h>
 
 #include "video_common.h"
-#include "video_device.h"
 
 LOG_MODULE_REGISTER(video_st_mipid02, CONFIG_VIDEO_LOG_LEVEL);
