+++ subsys/video/Kconfig
@@ -1,2 +1,35 @@
-# SPDX-FileCopyrightText: The Zephyr Project Contributors
+# Copyright The Zephyr Project Contributors
 # SPDX-License-Identifier: Apache-2.0
+
+menuconfig VIDEO
+	bool "Video subsystem"
+	select VIDEO_DRIVERS
+	help
+	  Enable support for the video subsystem as well as video drivers.
+
+if VIDEO
+
+module = VIDEO
+module-str = video
+source "subsys/logging/Kconfig.template.log_config"
+
+config VIDEO_SHELL
+	bool "Video shell"
+	depends on SHELL
+	depends on VIDEO
+	help
+	  This shell provides control and query commands for video drivers.
+
+if VIDEO_SHELL
+
+config VIDEO_SHELL_CTRL_NAME_SIZE
+	int "Maximum size for the control string identifier"
+	default 40
+	help
+	  Video controls have a human-friendly name that is converted into a string identifier.
+	  This Kconfig controls the maximum size of the string identifier after which the name
+	  would be truncated. The default value is enough for all standard control names.
+
+endif # VIDEO_SHELL
+
+endif # VIDEO
